Refactored StudyListLogic to fix obscure filtering bugs and ported old text query code
