+++ src/shared/components/query/studyList/styles.module.scss
